requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyterlab
snowflake-connector-python
python-dotenv